Add RedditResponse model to remain closer to the actual json
The interesting parts of the JSON result are wrapped in a data element.
We can unwrap this during deserialization in the ItemTypeAdapter or
follow the JSON with our data model.

With the second option we have a clear correspondence to the JSON but
we have to unwrap the result manually.